requirements.txt
beautifulsoup4
broth
date-extractor
python-docx
requests
openpyxl
validators
